refactor(repo): move multigateway to go/services/multigateway
Aligns directory structure with documented architecture. The multigateway
service code belongs in go/services/ per architecture.md. This resolves
the TODO comment in .golangci.yml.
