+++ go/services/multigateway/discovery.go
