+++ go/services/multigateway/discovery_test.go
