+++ go/services/multigateway/engine/plan.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/multigres/multigres/go/common/pgprotocol/server"
 	"github.com/multigres/multigres/go/common/sqltypes"
-	"github.com/multigres/multigres/go/multigateway/handler"
+	"github.com/multigres/multigres/go/services/multigateway/handler"
 )
 
 // Plan represents a query execution plan.
