+++ go/services/multigateway/engine/route.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/multigres/multigres/go/common/pgprotocol/server"
 	"github.com/multigres/multigres/go/common/sqltypes"
-	"github.com/multigres/multigres/go/multigateway/handler"
+	"github.com/multigres/multigres/go/services/multigateway/handler"
 )
 
 // Route is a primitive that routes a query to a specific tablegroup.
