+++ go/services/multigateway/planner/variable_set_stmt.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/multigres/multigres/go/common/parser/ast"
 	"github.com/multigres/multigres/go/common/pgprotocol/server"
-	"github.com/multigres/multigres/go/multigateway/engine"
+	"github.com/multigres/multigres/go/services/multigateway/engine"
 )
 
 // planVariableSetStmt plans SET/RESET commands.
